fix: Do not split fenced blocks by empty lines
If an opening fence is detected, let the chunker continue reporting
updates to the same block, until the closing fence is detected, when
the end of the block will be reported.

The flag `isCompleted` isn't reported by the parser any more, because
the unbalanced fences are needed by the chunker already. It appends
the flag `isCompleted` to both `block-update` and `block-end` events.

BREAKING CHANGE: This doesn't effect the usual usage of the `Semidown`
class - updating the container element with HTML markup rendered from
a Markdown stream. However, the interface of the inner objects was
updated. The parser returns only `html`, not `isComplete` any more.
The chunker adds `isComplete` to `block-update` and `block-end` event
payloads. The renderer accept `isComplete` in both `updateBlock` and
`finalizeBlock` methods.` Also, processing of fenced blocks changed
- they're constantly re-parsed and re-rendered until the closing fence
is detected. Although not a change on the API, the change of behaviour
can be considered breaking too.